requirements.txt
pysha3
construct
